fix: update styles for Article page and Categories
