+++ src/lib/components/ArticleMetadata.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import { type SubmitFunction, enhance } from '$app/forms';
-	import { type Article, ArticleStatus } from '$lib/articles';
+	import { type Article, ArticleSize, ArticleStatus } from '$lib/articles';
+	import type { Reactions } from '$lib/reactions';
 	import { Sentiment } from '$lib/utils';
 	import { formatDistance } from 'date-fns';
 
@@ -8,6 +9,7 @@
 
 	export let article: Article;
 	export let isActionable: boolean = false;
+	export let size: ArticleSize;
 
 	$: reactions = article.reactions;
 	$: mostPopularReaction = reactions.byType.sort((a, b) => b.total - a.total)[0].reaction;
@@ -35,6 +37,15 @@
 			await update();
 		};
 	};
+
+	const handleReaction: SubmitFunction = () => {
+		return async ({ result, update }) => {
+			if (result.type === 'success' && result.data) {
+				article = { ...article, reactions: result.data as Reactions };
+			}
+			await update();
+		};
+	};
 </script>
 
 <nav class="metadata">
@@ -47,6 +58,7 @@
 			})}
 		</time>
 	</a>
+
 	<div class="metadata__actions">
 		{#if !isDraft}
 			<a class="article-reactions-summary" href="/article/{article.id}">
@@ -88,6 +100,51 @@
 			</nav>
 		{/if}
 	</div>
+
+	{#if size === ArticleSize.FULL}
+		<nav class="article-reactions">
+			{#each article.reactions.byType as reaction}
+				<form
+					class="article-reactions__form"
+					action="/article/{article.id}?/react"
+					method="POST"
+					use:enhance={handleReaction}
+				>
+					<input type="hidden" name="reaction" value={reaction.index} />
+					<input type="hidden" name="article" value={article.id} />
+					<button
+						type="submit"
+						class="article-reactions__button
+							{article.reactions?.byCurrentUser === reaction.index ? 'article-reactions__button--reacted' : ''}"
+						disabled={!article.user}
+					>
+						<span class="article-reactions-summary__emoji">
+							{reaction.reaction}
+						</span>
+						{#if reaction.total}
+							<span class="article-reactions-summary__total">
+								{reaction.total}
+							</span>
+						{/if}
+					</button>
+				</form>
+			{/each}
+		</nav>
+	{/if}
+
+	{#if size === ArticleSize.FULL && article.messages}
+		<div class="article-prompt">
+			<code class="article-prompt__code">
+				{#each article.messages as message}
+					{#if typeof message.content === 'string'}
+						<p>{message.content}</p>
+					{:else if message.content?.notes !== undefined}
+						<p>{message.content.notes}</p>
+					{/if}
+				{/each}
+			</code>
+		</div>
+	{/if}
 </nav>
 
 <style lang="scss">
@@ -113,11 +170,11 @@
 
 	a.metadata__a {
 		display: flex;
+		gap: 2px;
 		flex-direction: column;
 		color: inherit;
 		text-decoration: none;
-		width: 100%;
-		height: 100%;
+		width: max-content;
 
 		&:hover {
 			color: var(--color-primary);
@@ -128,6 +185,8 @@
 		display: flex;
 		column-gap: 8px;
 	}
+
+	/* ------------------------------------------------------------------------ */
 
 	a.article-reactions-summary {
 		display: flex;
@@ -156,5 +215,67 @@
 
 	span.article-reactions-summary__total {
 		transform: translateY(1px); // Optically align with text `a.article-reactions-summary`
+	}
+
+	/* ------------------------------------------------------------------------ */
+
+	nav.article-reactions {
+		display: flex;
+		flex-direction: column;
+		width: max-content;
+		background-color: var(--color-neutral-700);
+		border-radius: var(--border-radius-l);
+		overflow: hidden; // Hide rounded corners
+	}
+
+	button.article-reactions__button {
+		display: flex;
+		align-items: center;
+		justify-content: center;
+		column-gap: 8px;
+		width: 100%;
+		padding: 8px;
+		box-sizing: border-box;
+		border: none;
+		background-color: transparent;
+		cursor: pointer;
+
+		// Styles shared with `a.article-reactions-summary`
+		text-align: center;
+		font-weight: 400;
+		line-height: 24px;
+		font-size: 12px;
+		font-family: var(--font-mono);
+		color: var(--color-neutral-200);
+
+		&--reacted {
+			font-weight: 600;
+		}
+
+		&--reacted,
+		&:hover:not(:disabled) {
+			background-color: var(--color-neutral-600);
+		}
+
+		&:disabled {
+			cursor: not-allowed;
+		}
+	}
+
+	/* ------------------------------------------------------------------------ */
+
+	div.article-prompt {
+		display: flex;
+		flex-direction: column;
+	}
+
+	code.article-prompt__code {
+		font-size: 13px;
+		font-family: var(--font-mono);
+		overflow-y: scroll;
+		color: var(--color-neutral-300);
+		background-color: var(--color-neutral-700);
+		border-radius: var(--border-radius-l);
+		padding: 0 16px;
 	}
 </style>
